requirements.txt
Flask==2.2.2
numpy==1.23.4
opencv-python==4.6.0.66
Pillow==9.2.0
requests==2.28.1
tensorflow==2.11.0
